refactor: :art: Rename components
