+++ components/NowPlaying.tsx
